Internal private header file with eager C struct definitions.
PiperOrigin-RevId: 171218337
